+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compiledgraph/CompiledParameter.scala
@@ -1,22 +1,31 @@
 package pl.touk.nussknacker.engine.compiledgraph
 
-import pl.touk.nussknacker.engine.api.definition.Parameter
+import pl.touk.nussknacker.engine.api.definition._
 import pl.touk.nussknacker.engine.api.expression.ExpressionTypingInfo
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.expression.parse.{CompiledExpression, TypedExpression}
+import pl.touk.nussknacker.engine.graph.expression.Expression
 
 object CompiledParameter {
 
   def apply(
       typedExpression: TypedExpression,
-      parameterDefinition: Parameter
+      parameterDefinition: Parameter,
+  ): CompiledParameter =
+    apply(typedExpression, parameterDefinition, Nil)
+
+  def apply(
+      typedExpression: TypedExpression,
+      parameterDefinition: Parameter,
+      validators: List[Validator],
   ): CompiledParameter = {
     CompiledParameter(
       parameterDefinition.name,
       typedExpression.expression,
       parameterDefinition.scalaOptionParameter,
       parameterDefinition.javaOptionalParameter,
-      typedExpression.typingInfo
+      typedExpression.typingInfo,
+      validators,
     )
   }
 
@@ -27,8 +36,20 @@ final case class CompiledParameter(
     override val expression: CompiledExpression,
     override val shouldBeWrappedWithScalaOption: Boolean,
     override val shouldBeWrappedWithJavaOptional: Boolean,
-    typingInfo: ExpressionTypingInfo
-) extends BaseCompiledParameter
+    typingInfo: ExpressionTypingInfo,
+    validators: List[Validator] = Nil,
+) extends BaseCompiledParameter {
+
+  lazy val expressionForValidation: Expression =
+    Expression(expression.language, expression.original)
+
+  private lazy val resolvedValidators: List[Validator] =
+    ParameterValidator.resolveLoaders(validators)
+
+  lazy val runtimeValidators: List[RuntimeValidator] =
+    resolvedValidators.collect { case v: RuntimeValidator => v }
+
+}
 
 trait BaseCompiledParameter {
   def name: ParameterName